refactor: add tenant controller

+++ src/modules/core/kv/kv.controller.ts
@@ -1,9 +1,9 @@
-import { Body, Controller, Get, Post, Query, UseGuards } from '@nestjs/common';
+import { Body, Controller, Get, Post, Query, Req, UseGuards } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 import { IsOptional, IsString } from 'class-validator';
 import { r } from '../../common/helpers';
 import { LoggerFactory } from '../../common/logger';
-import { JwtAdminAuthGuard } from '../auth';
+import { AnyAuthGuard, AnyAuthRequest, JwtAdminAuthGuard } from '../auth';
 import { KeyValuePair, ValueType } from './kv.entities';
 import { KvDef, KvDefIdentifierHelper, KvHelper } from './kv.helper';
 
@@ -30,6 +30,10 @@ class KvPair {
   @IsString()
   @IsOptional()
   extra?: any;
+
+  toKvDef(): KvDef {
+    return { collection: this.collection, key: this.key };
+  }
 }
 
 class GetKvPairRequest {
@@ -43,38 +47,52 @@ class GetKvPairRequest {
   @IsString()
   @IsOptional()
   transform?: string;
+
+  toKvDef(): KvDef {
+    return { collection: this.collection, key: this.key };
+  }
 }
 
 @ApiTags('core')
 @Controller('api')
 export class KvController {
-  @UseGuards(new JwtAdminAuthGuard())
+  @UseGuards(JwtAdminAuthGuard)
   @Post('kv')
-  async set(@Body() kvPair: KvPair) {
-    logger.log(`set ${r(kvPair)}`);
+  async set(@Body() kvPair: KvPair, @Req() req: AnyAuthRequest): Promise<KeyValuePair> {
+    const { user, identifier } = req;
+    logger.log(`set ${r({ kvPair, user, identifier })}`);
+    await KvHelper.checkPermission(kvPair.toKvDef(), identifier);
     return KvHelper.set(KeyValuePair.create(kvPair));
   }
 
-  @UseGuards(new JwtAdminAuthGuard())
+  @UseGuards(JwtAdminAuthGuard)
   @Post('kv/destroy')
-  async destroy(@Body() kvDef: KvDef) {
-    logger.log(`destroy ${r(kvDef)}`);
+  async destroy(@Body() kvDef: KvDef, @Req() req: AnyAuthRequest): Promise<void> {
+    const { user, identifier } = req;
+    logger.log(`destroy ${r({ kvDef, user, identifier })}`);
+    await KvHelper.checkPermission(kvDef, identifier);
     await KvHelper.delete(kvDef);
     const initializer = KvHelper.initializers[KvDefIdentifierHelper.stringify(kvDef)];
-    initializer && (await initializer());
+    if (initializer) await initializer();
     return null;
   }
 
   // fixme kv 的读取需要一定的权限
+  @UseGuards(AnyAuthGuard)
   @Get('kv')
-  async get(@Query() query: GetKvPairRequest) {
-    logger.log(`get ${r(query)}`);
+  async get(@Query() query: GetKvPairRequest, @Req() req: AnyAuthRequest): Promise<KeyValuePair> {
+    const { user, identifier } = req;
+    logger.log(`get ${r({ query, user, identifier })}`);
+    await KvHelper.checkPermission(query.toKvDef(), identifier);
     return KvHelper.get(query.collection, query.key);
   }
 
+  @UseGuards(AnyAuthGuard)
   @Get('kvs')
-  collection(@Query('collection') collection: string) {
-    logger.log(`get kvs by ${collection}`);
+  async collection(@Query('collection') collection: string, @Req() req: AnyAuthRequest): Promise<KeyValuePair[]> {
+    const { user, identifier } = req;
+    logger.log(`get kvs by ${r({ collection, user, identifier })}`);
+    await KvHelper.checkPermission({ collection }, identifier);
     return KvHelper.find(collection);
   }
 }
